fix: tokenization + search weights
Looked at searches like '860', 'generators', 'generator' to determine
that we should have some more robust tokenization that splits e.g.
'eia860' into 'eia' and '860'. Also we should have stemming, but with
some special cases (without special cases, 'generator' and 'generation'
go to the same search term, which is useful in most cases but explicitly
not in our use case). This allows some fuzzier matching in the search.
Finally, fixed some of the search weights so that the out and
leading-underscore tables are properly boosted / penalized.